+++ Kernel/TTY/MasterPTY.cpp
@@ -5,13 +5,13 @@
  */
 
 #include <Kernel/API/POSIX/errno.h>
+#include <Kernel/API/POSIX/signal_numbers.h>
 #include <Kernel/Debug.h>
 #include <Kernel/InterruptDisabler.h>
 #include <Kernel/Process.h>
 #include <Kernel/TTY/MasterPTY.h>
 #include <Kernel/TTY/PTYMultiplexer.h>
 #include <Kernel/TTY/SlavePTY.h>
-#include <LibC/signal_numbers.h>
 #include <LibC/sys/ioctl_numbers.h>
 
 namespace Kernel {
