+++ Kernel/Thread.h
@@ -16,6 +16,7 @@
 #include <AK/Variant.h>
 #include <AK/Vector.h>
 #include <Kernel/API/POSIX/sched.h>
+#include <Kernel/API/POSIX/signal_numbers.h>
 #include <Kernel/Arch/RegisterState.h>
 #include <Kernel/Arch/ThreadRegisters.h>
 #include <Kernel/Debug.h>
@@ -31,7 +32,6 @@
 #include <Kernel/Memory/VirtualRange.h>
 #include <Kernel/UnixTypes.h>
 #include <LibC/fd_set.h>
-#include <LibC/signal_numbers.h>
 
 namespace Kernel {
