+++ Userland/Libraries/LibC/signal.h
@@ -9,7 +9,6 @@
 #include <Kernel/API/POSIX/signal.h>
 #include <Kernel/API/POSIX/ucontext.h>
 #include <bits/sighow.h>
-#include <signal_numbers.h>
 #include <sys/cdefs.h>
 #include <sys/types.h>
